GEODE-3253: Refactoring ClientCommands and related tests
